% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedVariableFeaturePlot.R
\name{AnnotatedVariableFeaturePlot}
\alias{AnnotatedVariableFeaturePlot}
\title{Annotated variable features plot}
\usage{
AnnotatedVariableFeaturePlot(
  object,
  which.labels = NA,
  pt.size = 1,
  text.size = 2.5,
  log = NULL,
  xnudge = 0,
  ynudge = 0,
  colors = c("Black", "Cyan"),
  selection.method = NULL,
  assay = NULL
)
}
\arguments{
\item{object}{Seurat object. Must contain nFeature_RNA and percent.mito metadata columns, or the x and y plot parameters must be changed accordingly to plot other QC info otherwise.}

\item{which.labels}{Number of most variable features to label, or list of variable features to label. If NA, label all the variable features (Default: NA).}

\item{pt.size}{Size of the points on the plot (Default: 1)}

\item{text.size}{Size of the feature annotations (Default: 2.5)}

\item{log}{Plot the x-axis on log scale}

\item{xnudge}{Offset in x on label position (Default: 0)}

\item{ynudge}{Offset in y on label position (Default: 0)}

\item{colors}{Colors to specify non-variable/variable status (Default: c("Black","Cyan")).}

\item{selection.method}{Which method to pull (refer to ?Seurat::VariableFeaturePlot()).}

\item{assay}{Assay to pull variable features from}
}
\value{
A ggplot object.
}
\description{
This function displays a variable feature plot with feature annotations.
}
\examples{
SeuratObject <- FindVariableFeatures(SeuratObject)
# Annotate all variable features:
AnnotatedVariableFeaturePlot(SeuratObject)
# Annotate user-defined genes:
AnnotatedVariableFeaturePlot(SeuratObject,c("OLFM4","LGR5","CD44","GAPDH"))
AnnotatedVariableFeaturePlot(SeuratObject,500)
}
\keyword{QC}
\keyword{Variable}
\keyword{annotated}
\keyword{annotations}
\keyword{features}
\keyword{genes}
\keyword{plot}
