% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateWithTCGA.R
\name{AnnotateWithTCGA}
\alias{AnnotateWithTCGA}
\title{Annotate with TCGA}
\usage{
AnnotateWithTCGA(genotype, tcga)
}
\arguments{
\item{genotype}{A genotype object}

\item{tcga}{A data frame that must contain "gene_ensembl_id" and "tcga_percent" columns (character and numeric respectively).}
}
\value{
A data frame with at least gene_ensembl_id, cases_in_cohort and tcga_percent columns.
}
\description{
Add TCGA mutation prevalence within a cohort data to the metadata of a genotype object.
}
\details{
See also ReadTCGA and https://portal.gdc.cancer.gov/exploration?searchTableTab=genes
}
\examples{
MyGenotypes <- ReadVcf("MyGenotypes.vcf.gz")
MyTCGA <- ReadTCGA("MyTCGAmutations.tsv")
MyGenotypes <- AnnotateWithTCGA(MyGenotypes,MyTCGA)
}
\keyword{alt}
\keyword{consensus}
\keyword{freq}
\keyword{frequency}
\keyword{matrix}
\keyword{ref}
\keyword{variants}
\keyword{vartrix}
