% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateWithVEP.R
\name{AnnotateWithVEP}
\alias{AnnotateWithVEP}
\title{Annotate variants with VEP data.}
\usage{
AnnotateWithVEP(
  genotype,
  vep,
  min_impact = "MODERATE",
  avoid_underscores = TRUE
)
}
\arguments{
\item{genotype}{A genotype object, typically from ReadVcf.}

\item{vep}{A data frame or tibble, typically from ReadVEP, containing the fields "Uploaded_variation", "Location", "Allele", "Gene", "SYMBOL", "IMPACT", "Consequence", "Protein_position", "Amino_acids", "CLIN_SIG"}

\item{min_impact}{character(1). Minimal impact included in the short descriptions of variants, among: "-"<"MODIFIER"<"LOW"<"MODERATE"<"HIGH". Default: "MODERATE"}

\item{avoid_underscores}{logical. Should underscores be replaced by dashes as a field separator. Needed when the variant data will be used with Seurat, which replaces underscore in feature names by dashes. Default: TRUE.}
}
\value{
A genotype object with additional metadata fields (symbol gene_ensembl_id impact protein_position aminoacids clinicalsign consequence summary short_description) and annotated variant names (symbol-protein_position-aminoacids-impact-chr_pos-ref-alt).
}
\description{
Add variant effect predictor (VEP) annotations to a genotype object, and renames the variants with a short human-intelligible description of the variant.
}
\details{
If available in the VEP data frame, adds the fields: symbol, gene_ensembl_id, impact, protein_position, aminoacids, clinicalsign, summary, and- short_description to the genotype object metadata. Also renames the variants (typically named chr-pos-ref-alt after reading a vcf) with a short description prefix of the form symbol-protein_position-aminoacids-impact. Fields for which no data is available, or impact values less than min_impact, are skipped from the short description.

VEP typically assigns several annotations for each variant in a vcf file. The one selected for annotating the genotype object is the one with the highest predicted impact. If several have the same impact, the ones with a symbol available are prioritized, and among those the ones with a protein position / amino acid change prediction. If there is still a draw between several annotations, the first one listed in the original VEP data is used.

refers to HGNC gene symbols.
}
\examples{
MyGenotype <- AnnotateWithVEP(MyGenotype,MyVepDataFrame,min.impact="HIGH",avoid_underscores=TRUE)
}
\keyword{annotate}
\keyword{genotype}
\keyword{variants}
\keyword{vep}
