% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyManual.R
\name{ClassifyManual}
\alias{ClassifyManual}
\title{Manual cell type classification based on signature gating}
\usage{
ClassifyManual(object, gates, metadata.name = "celltype")
}
\arguments{
\item{object}{Seurat object. Must have metadata columns containing signature scores, such as those generated by ScoreSignatures, with names matching any gating filter supplied.}

\item{gates}{A string (chr) of cell type definitions, with cell types delimited by commas (,). The name of the cell type is given first, followed by equal (=) and a series of gates separated by AND symbol (&). Example: "celltype_a = sign_a > value_aa & sign_b < value_ab, celltype_b = sign_b > value_ba & sign_a < value_ab" where sign_a and sign_b are the names of metadata columns and value_xy are numeric values.}

\item{metadata.name}{The name of the new metadata column where cell type annotations are stored (Default: celltype)}
}
\value{
A Seurat object with an additional metadata column containing the cell type annotations and Idents() set to these annotations. Cells which do not validate any gate defined are attributed the celltype "None" and cells which pass several gates are labelled "Multiplet".
}
\description{
This function performs a cell type classification based on signature gating.
}
\details{
If a list of signatures (sign_a, sign_b...) has been scored with ScoreSignature, cell types (celltype_a, celltype_b...) are defined based on gates (value_xy cutoffs) of the form:
"celltype_a = sign_a > value_aa & sign_b < value_ab, celltype_b = sign_b > value_ba & sign_a < value_ab"
}
\examples{
# After MySeuratObject has been log-normalized in order to contain an RNA>data assay>slot and been augmented with a percent.mito metadata column:
# SignatureList should be a list of character vectors, each containing a series of feature/gene names. In this example, the signatures correspond to cell types, and the name of each signature is the name of the corresponding cell type (e.g. Tcell, Bcell, EpithelialCell, Fibroblasts, etc).
MySeuratObject <- Impute(MySeuratObject)
MySeuratObject <- ScoreSignatures(MySeuratObject,SignatureList)
gates <- "celltype_a = sign_a > 1.2 & sign_b < 0.2,
          celltype_b = sign_b > 1.1 & sign_a < 0.3"
MySeuratObject <- ClassifyManual(MySeuratObject,gates)
}
\keyword{Cell}
\keyword{Classifier}
\keyword{celltype}
\keyword{classification}
\keyword{gates}
\keyword{gating}
\keyword{manual}
\keyword{type}
