% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMitoContent.R
\name{ComputeMitoContent}
\alias{ComputeMitoContent}
\title{Compute the fraction/percentage of mitochondrial genes (mouse and human) in Seurat objects}
\usage{
ComputeMitoContent(object, name = "percent.mito", as.percent = FALSE)
}
\arguments{
\item{object}{Seurat object. Must contain an RNA assay.}

\item{name}{Name of the new metadata column created containing the fraction of mitochondrial transcripts (Default: percent.mito)}

\item{as.percent}{Whether to report the result as a fraction (between 0 and 1) or percentage (between 0 and 100). Default: TRUE (fraction).}
}
\value{
Seurat object with an additional metadata column containing mitochondrial fraction/percentage.
}
\description{
This function computes the mitochondrial content of cells in a Seurat object and adds it to the metadata.
}
\examples{
MySeuratObject <- ComputeMitoContent(MySeuratObject)
}
\keyword{Mito}
\keyword{Mitochondrial}
\keyword{content}
\keyword{fraction}
\keyword{genes}
\keyword{percentage}
