% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossImpute.R
\name{CrossImpute}
\alias{CrossImpute}
\title{MAGIC imputation of features of an assay based on the distances in another assay}
\usage{
CrossImpute(
  object,
  impute.assay = "default_assay",
  name = "crossimputed",
  impute.features = "all",
  reference.assay = "RNA",
  reference.features = "variable_features",
  npca = 40,
  knn = 3,
  t = 2,
  n.jobs = 6,
  slot.use = "data"
)
}
\arguments{
\item{object}{Seurat object.}

\item{impute.assay}{character(1). The name of the assay which should be imputed, or "metadata", or "default_assay" (Default: "default_assay").}

\item{name}{character(1). The name under which the resulting imputed assay is stored in the Seurat object (Default: "crossimputed")}

\item{impute.features}{character(n). Features on which to run the imputation in the impute.assay, or "all" for all features in the impute.assay, or "variable_features" for the variable features in the impute.assay (Default:"all").}

\item{reference.assay}{character(1). The name of the assay which is appended to the impute.assay to guide imputation distances.}

\item{reference.features}{character(n). Features from the reference.assay to use to guide the imputation of the impute.assay, or "all" for all features in the reference.assay, or "variable_features" for the variable features in the reference.assay (Default:"variable_features").}

\item{npca}{integer(1). Number of principal components to use for imputation.}

\item{knn}{integer(1). Number of nearest neighbors on which to compute bandiwth imputation.}

\item{t}{integer(1). Diffusion parameter passed to MAGIC.}

\item{n.jobs}{integer(1). Number of threads on which to run the imputation (Default: 6).}

\item{slot.use}{character(1). Name of the slot to impute (Default: "data").}
}
\value{
Seurat object with an additional slot containing MAGIC cross-imputed data.
}
\description{
This function imputes features in an assay based on (the merger with) another assay. Typically using the high resolution in the RNA data to impute more noisy additional assays.
}
\examples{
MySeuratObject <- CrossImpute(MySeuratObject,"other_assay_to_impute_based_on_RNA_distances")
MySeuratObject[["crossimputed"]] # Check result
}
\keyword{MAGIC}
\keyword{Rmagic}
\keyword{cross-imputation}
\keyword{crossimputation}
\keyword{imputation}
