% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportGenotype.R
\name{ExportGenotype}
\alias{ExportGenotype}
\title{Reexport vartrix matrices from a genotype object}
\usage{
ExportGenotype(
  genotype,
  dir,
  meta_columns = "all",
  append_genotypes = T,
  gzip = T,
  rows = "vartrix_variants.tsv",
  cols = "vartrix_barcodes.tsv",
  ref_counts = "vartrix_ref_matrix.mtx",
  alt_counts = "vartrix_alt_matrix.mtx",
  meta = "vartrix_metadata.tsv"
)
}
\arguments{
\item{genotype}{A genotype object}

\item{dir}{character(1). The directory in which the data should be exported. Created if non-existent.}

\item{meta_columns}{character(n). Metadata columns that should be exported from the genotype Object. Can also be "all", or empty vector c() for none. Default: "all".}

\item{append_genotypes}{logical(1). Whether the bulk genotypes should also be exported (appended to metadata). Default: TRUE.}

\item{gzip}{logical(1). Should the exported files be compressed. If compressed, the filenames are appended with the extension ".gz". Default: TRUE}

\item{rows}{character(1). Name of the exported file that lists row names. Default: vartrix_variants.tsv}

\item{cols}{character(1). Name of the exported file that lists column names. Default: vartrix_barcodes.tsv}

\item{ref_counts}{character(1). Name of the exported file that contains the reference allele sparse raw count matrix. Default: vartrix_ref_matrix.mtx}

\item{alt_counts}{character(1). Name of the exported file that contains the alternate allele sparse raw count matrix. Default: vartrix_alt_matrix.mtx}

\item{meta}{character(1). Name of the exported file that contains the metadata. Default: metadata.tsv}
}
\description{
Exports the reference and alternate count matrices, as well as the barcodes (colnames), variant names (rownames) and metadata/reference genotypes.
}
\examples{
# By default keeping all the metadata columns, and no dimensionality reductions.
Export10X(SeuratObject, "MyDir")
# Include only two custom metadata columns, two dimensionality reductions, and do not compress the matrices.
Export10X(SeuratObject, "MyDir", c("nFeature_RNA", "mito.content"), c("pca","umap"), gzip=FALSE)
}
\keyword{Reexport}
\keyword{export}
\keyword{genotype}
\keyword{variants}
