% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeSimilarityHeatmap.R
\name{GenotypeSimilarityHeatmap}
\alias{GenotypeSimilarityHeatmap}
\title{Genotype similarity heatmap}
\usage{
GenotypeSimilarityHeatmap(
  genotypes,
  key.title = "Fraction of \\nmatching variants",
  limits = NULL,
  plot_method = "plotly",
  colors = viridis::viridis(n = 256, alpha = 1, begin = 0, end = 1, option = "viridis"),
  ...
)
}
\arguments{
\item{genotypes}{genotype object, contains several genotypes to be compared on the heatmap.}

\item{key.title}{character(1), the legend/key. Default: "Fraction of matching variants"}

\item{limits}{numeric(2), the limits of the color scale (typically c(0,1), which are the extremes values similarities could take), or NULL for min/max in current data.}

\item{plot_method}{"plotly" or "ggplot". Default "plotly"}

\item{colors}{a color palette, default viridis.}

\item{...}{other graphic parameters passed on to heatmaply.}
}
\value{
A heatmaply plot.
}
\description{
Plot a heatmap of the similarity between the various genotypes in a genotype object.
}
\details{
The similarity is defined as the fraction of matching variants, i.e. the number of identical variants divided by the number of variants which are simultaneously called in both genotypes being compared.
}
\examples{
GenotypeSimilarityHeatmap(MyGenotypeObject)
}
\keyword{genotypes}
\keyword{variants}
\keyword{vcf}
