% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllSignatureGenes.R
\name{GetAllSignatureGenes}
\alias{GetAllSignatureGenes}
\title{Get all signature genes}
\usage{
GetAllSignatureGenes(signature.list)
}
\arguments{
\item{signature.list}{Character vector list. A list of signatures.}
}
\value{
The unique elements present among the signatures (character vector).
}
\description{
This function returns all the unique genes appearing within a list of signatures. Particularly useful when signatures contain "negative" genes, as it will remove the "-" sign. Also works on single vectors rather than lists, in which case it will remove the minus sign and duplicates.
}
\examples{
SL <- list(CD8TC=c("CD3","CD8","-CD4"), 
           Fibro=c("VIM","COL1A1","-EPCAM"))
UniqueGenes <- GetAllSignatureGenes(SL) # Returns c("CD3","CD4","CD8","VIM","COL1A1","EPCAM")
}
\keyword{Unique}
\keyword{genes}
\keyword{signatures}
