% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentPlotQC.R
\name{IdentPlotQC}
\alias{IdentPlotQC}
\title{QC plot with cell type highlight}
\usage{
IdentPlotQC(
  object,
  ident = NA,
  x = "nFeature_RNA",
  y = "percent.mito",
  log.scale = TRUE,
  ncol = NA,
  pt.size = 1,
  interactive = TRUE,
  colors.use = NULL,
  assay = DefaultAssay(object),
  slot = "data",
  ...
)
}
\arguments{
\item{object}{Seurat object. Must contain nFeature_RNA and percent.mito metadata columns, or the x and y plot parameters must be changed accordingly to plot other QC info otherwise.}

\item{ident}{The name of the categorical metadata column to use as cell types (Default: NA, using the current active annotations, i.e. Idents(object)). If ident is a list, an array of plots for the various set of annotations provided will be returned.}

\item{x}{Which parameter to set to the x axis in the QC plot (Default: nFeature_RNA).}

\item{y}{Which parameter to set to the y axis in the QC plot (Default: percent.mito).}

\item{log.scale}{Whether to plot with log.scale or not (Default: TRUE)}

\item{ncol}{Number of columns (Default = NA, determined from the data)}

\item{pt.size}{The point size, passed to ggplot (Default: 1)}

\item{interactive}{Enable interactive plot? (Default: TRUE).}

\item{colors.use}{For interactive plots, a colorbrewer2.org palette name (e.g. "YlOrRd" or "Blues"), or a vector of colors to interpolate in hexadecimal "#RRGGBB" format, or a color interpolation function like colorRamp(). For non-interactive plots, a vector of colors passed to ggplot2 scale_color_manual. If the vector is named, the values will be matched based on names.}

\item{assay}{character(1). If axes are features rather than metadata columns, the assay from which they should be pulled in priority (Default: DefaultAssay(object)).}

\item{slot}{character(1). If axes are features rather than metadata columns, the slot from which they should be pulled (Default: "data").}

\item{...}{Other arguments passed to plot_ly in the case of interactive plots.}
}
\value{
If one ident is plotted and interactive is enabled, returns interactive plot (plotly). If several, returns a ggplot grid (cowplot).
}
\description{
This function makes interactive quality control (QC) plots from a Seurat object, by default percent.mito vs nFeature_RNA, while color-coding various cell types. Any metadata column, signature, or feature can be also be used for the x/y axis. To pick from a particular assay, use an underscore separator, as in "assay_feature" or set the assay parameter. Features are picked up from the data slot by default, set slot to change.
}
\details{
Plotly interactions: click on legend categories to toggle the display on/off, double click to isolate one cell type/category/ident.
Hover to get cell info (useful to choose filtering parameters), and select pan/zoom in the top-right menu to do close-ups and navigate.

This is useful when making scRNA-seq from heterogeneous samples: for example immune cells are small and might be discarded as debris from much larger epithelial cancer cells if one is not aware of the differences in QC parameters occuring between various cell types.
}
\examples{
# After MySeuratObject has been log-normalized in order to contain an RNA>data assay>slot and been augmented with a percent.mito metadata column:
# SignatureList should be a list of character vectors, each containing a series of feature/gene names. In this example, the signatures correspond to cell types, and the name of each signature is the name of the corresponding cell type (e.g. Tcell, Bcell, EpithelialCell, Fibroblasts, etc).
MySeuratObject <- Impute(MySeuratObject)
MySeuratObject <- ScoreSignatures(MySeuratObject,SignatureList)
MySeuratObject <- Classify(MySeuratObject,names(SignatureList),"CellType") # Automatic cell type annotation based on cell type signatures.
IdentPlotQC(MySeuratObject,"celltype")
}
\keyword{QC}
\keyword{cell}
\keyword{highlighted.}
\keyword{plot}
\keyword{types}
\keyword{with}
