% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformativeVariantPlot.R
\name{InformativeVariantPlot}
\alias{InformativeVariantPlot}
\title{Plot the informative variants with annotations}
\usage{
InformativeVariantPlot(
  genotype,
  do.label = T,
  label.column = "variants",
  label.only.informative = T,
  min.impact = NA,
  min.tcga.percent = NA,
  check.overlap = T,
  pt.size = 1,
  text.size = 3.5,
  colors.use = c("grey", "blue", "red"),
  log.scale = F,
  min.entropy = 0.001
)
}
\arguments{
\item{genotype}{A genotype object}

\item{do.label}{logical(1). Should variants be labeled (Default: TRUE).}

\item{label.column}{character(1). Which metadata column in the genotype object should be used as labels (Default: variants).}

\item{label.only.informative}{logical(1). Should the labels only be for informative variants (Default: TRUE).}

\item{min.impact}{character(1). Label only variants which have a predicted impact more than "LOW", "MODERATE", or "HIGH". NA to label all (Default:NA).}

\item{min.tcga.percent}{character(1). Label only variants which concern genes mutated in more than a certain percent of TCGA patients in the cohort used for genotype annotation. Requires a tcga_percent column to be present in the genotype object.}

\item{check.overlap}{logical(1). Should overlapping labels be omitted (Default: TRUE).}

\item{pt.size}{numeric(1). The point size (Default: 1).}

\item{text.size}{numeric(1). The text size (Default:3.5).}

\item{colors.use}{character(3). The colors for other variants, informative variants, and high impact / high tcga percent variants, in this order (Default: c("grey","blue","red")).}

\item{log.scale}{logical(1). Should the axis be switched to log scale (Default: FALSE).}

\item{min.entropy}{numeric(1). Minimal excess_entropy value for variants to be represented by a dot in the graph. Values above 0, even small, greatly speed up rendering without noticeably changing the result, if many very low informative variants are present. More critical when using log scale, as low entropy variants are then visible (Default:1e-3).}
}
\value{
Returns a ggplot
}
\description{
Plots the excess entropy vs coverage of the variants, highlighting in color the informative variants. Optionally, can label all or only informative variants, or the variants having a minimal impact or present in a minimal percent of TCGA patients for the relevant cohort, or a combination thereof.
}
\examples{
InformativeVariantPlot(MyGenotypes,label.column = "summary") # Other labels that can be interesting: symbol, summary, short_description.. see also colnames(MyGenotypes@metadata)
InformativeVariantPlot(MyGenotypes,label.column = "summary",min.impact = "MODERATE") # Only label the variants that have a predicted impact more than moderate
InformativeVariantPlot(MyGenotypes,label.column = "summary",min.impact = "MODERATE",min.tcga.percent = 1,label.only.informative=F,log.scale = T) # Add the restriction to only label above a minimal tcga percent. The results are few, so label even non-informative variants, and activate log scale to see better the low coverage variants.
}
\keyword{annotate}
\keyword{genotype}
\keyword{variants}
\keyword{vep}
