% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeHTO.R
\name{NormalizeHTO}
\alias{NormalizeHTO}
\title{Normalize hashtag oligo data}
\usage{
NormalizeHTO(
  object,
  scale.HTO = "stdev",
  normalize.cells = 100,
  assay = "HTO",
  q = 0.9,
  knn = 50,
  t = 10
)
}
\arguments{
\item{object}{Seurat object.}

\item{scale.HTO}{character(1) or logical(1). "stdev"/TRUE/T to scale each HTO to standard deviation across cells, No/no/FALSE/F for no scaling, "quantile" to scale to a quantile of values, and "imputed_max" to scale to the max after smoothing by imputation to avoid outliers. Default: "stdev".}

\item{normalize.cells}{numeric(1). The value to which each cell should be normalized. 0 for no cell-normalization. Default: 100.}

\item{assay}{character(1). The name of the assay which should be normalized. Default: "HTO".}

\item{q}{numeric(1). Only used if scale.HTO="quantile". The quantile used for normalization.}

\item{knn}{numeric(1). Only used if scale.HTO="imputed_max". Passed to CrossImpute.}

\item{t}{numeric(1). Only used if scale.HTO="imputed_max". Passed to CrossImpute.}
}
\value{
Returns a Seurat object with normalized HTO data in the data slot.
}
\description{
Normalizes hashtag oligo (HTO) data. By default, first scales each HTO to its standard deviation to account for difference of intensity for the various labels, then scales each cell so that HTO are reported as a percentage. Alternatively, can scale the HTOs to a quantile of their values, or to their max after imputation.
}
\examples{
MySeuratObject <- NormalizeHTO(MySeuratObject)
}
\keyword{HTO}
\keyword{normalization}
