% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefilterVartrix.R
\name{PrefilterVartrix}
\alias{PrefilterVartrix}
\title{Prefilter variants by coverage in the vartrix matrices}
\usage{
PrefilterVartrix(genotype, min.cells = 5, barcodes.keep = NA)
}
\arguments{
\item{genotype}{A genotype object, containing vartrix matrices (including a VAR consensus matrix, named as such, used to compute coverage).}

\item{min.cells}{numeric(1) the minimum number of cells for which there should be a genotype call for the variants to be kept.}

\item{barcodes.keep}{character(n) A vector of cell barcodes to keep. NA to keep all. Default:NA.}
}
\value{
The filtered genotype object
}
\description{
Given a genotype object containing a list of vartrix matrices, and a minimal number of cells/barcodes for which variant data should be available, returns a list of vartrix matrices with filtered variants. If barcodes are given, start by restricting the matrices to these barcodes before filtering variants.
The row order and rownames are kept consistent with the input during the subsetting. The subsetting is also applied to the entire genotype object, including reference genotypes and variant metadata. Stores the coverage information in genotype$coverage.
}
\examples{
MyFilteredGenotypeObject <- PrefilterVartrix(MyGenotypeObject, min.cells=10)
}
\keyword{genotypes}
\keyword{variants}
\keyword{vcf}
