% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulkGenotypes.R
\name{PseudoBulkGenotypes}
\alias{PseudoBulkGenotypes}
\title{Aggregate pseudo-bulk genotypes}
\usage{
PseudoBulkGenotypes(
  object,
  ident = Idents(object),
  tolerance.pct = 5,
  assays = c("VAR", "REF", "ALT"),
  layer = "data"
)
}
\arguments{
\item{object}{Seurat object.}

\item{ident}{character(n). The name of a metadata column, or a vector listing the individual identities of cells, or a list of cells. Defines the groups of cells that are aggregated in pseudo-bulk genotypes. Default: Idents(object).}

\item{tolerance.pct}{numeric(1). The percentage of counts not matching the rest tolerated to keep a homozygous call.}

\item{assays}{character(3). The name of the assays in which the consensus matrix (expected to also contain variant metadata), reference allele count and alt allele count are stored in the seurat object. Default: c("VAR","REF","ALT").}

\item{layer}{character(1). The assay layer that should be used in the Seurat object. Default: "data".}
}
\value{
Returns a genotype object with the aggregated pseudo-bulk genotypes and their metadata.
}
\description{
Starting from a Seurat object, for each of the identities given as the ident argument, compute a pseudo-bulk aggregated genotype. Store the results in a genotype object. If the ident is a list of cells, compute only one aggregated genotype for the cells listed. If it is the name of a metadata column, or a vector listing the identity of each individual cells, use the levels of this vector/column to define the groups of cells aggregated into individual pseudo-bulk genotypes.
}
\examples{
GenotypeObject <- PseudoBulkGenotypes(MySeuratObject)
}
\keyword{Pseudobulk}
\keyword{aggregation}
\keyword{genotype}
\keyword{pseudo-bulk}
