% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTCGA.R
\name{ReadTCGA}
\alias{ReadTCGA}
\title{Load TCGA data}
\usage{
ReadTCGA(file)
}
\arguments{
\item{file}{character(1), the tsv file path.}
}
\value{
A data frame with at least gene_ensembl_id, cases_in_cohort and tcga_percent columns.
}
\description{
Read a tsv file with TCGA most commonly mutated genes for a given cohort.
}
\details{
Read a tab separated value (tsv) file with most commonly mutated genes for a given cancer cohort according to the cancer genome atlas (TCGA). The tsv file should have columns with Simple Somatic Mutation (SSM) Affected Cases in cohort, used to derive a percentage of patients affected, as well as gene (ensembl) IDs. This import function follows the conventions of files exported in tsv from:

https://portal.gdc.cancer.gov/exploration?searchTableTab=genes
}
\examples{
tcga_df <- ReadTCGA("Myfolder/MyTCGAcohortMostCommonMutatedGenes.tsv")
}
\keyword{TCGA}
\keyword{annotation}
\keyword{atlas}
\keyword{cancer}
\keyword{common}
\keyword{genome}
\keyword{mutations}
\keyword{variants}
