% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVcf.R
\name{ReadVcf}
\alias{ReadVcf}
\title{Read a vcf file}
\usage{
ReadVcf(file = "")
}
\arguments{
\item{file}{The path to the vcf file, can be compressed (.vcf or .vcf.gz).}
}
\value{
genotype object
}
\description{
This function reads a vcf file and returns the contents as a genotype object.
}
\details{
This function reads a vcf file, containing the tab separated fields CHROM, POS, REF, ALT, ID, QUAL, INFO, FORMAT, and an undefined additional number of columns with genotypes (column titles are the genotype names, typically patient IDs), and returns a genotype object (genotypes as a sparse matrix in vartrix conventions, and associated metadata as a data frame).
}
\examples{
genotypes <- ReadVcf("MyFolder/Myfile.vcf.gz")
}
\keyword{genotypes}
\keyword{variants}
\keyword{vcf}
