% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQC.R
\name{RunQC}
\alias{RunQC}
\title{QC by cell type}
\usage{
RunQC(object, celltypes, parameter, min.val = 0, max.val = Inf)
}
\arguments{
\item{object}{Seurat object. Must contain nFeature_RNA and percent.mito metadata columns, or the x and y plot parameters must be changed accordingly to plot other QC info otherwise.}

\item{celltypes}{Identities that should be filtered. Must be present in Idents(object).}

\item{parameter}{Metadata column on which to do the filtering.}

\item{min.val}{Minimal accepted values for each identity (Default: 0).}

\item{max.val}{Maximal accepted values for each identity (Default: +Inf).}
}
\value{
A Seurat object
}
\description{
This function does cell-type aware quality control for scRNA-seq data in Seurat objects. Cells listed in 'celltypes' are filtered, keeping only cells for which 'parameter' is between min.val and max.val.
}
\examples{
MySeuratObject <- RunQC(MySeuratObject)
MySeuratObject <- ScoreSignatures(MySeuratObject,SignatureList)
MySeuratObject <- Classify(MySeuratObject,names(SignatureList),"CellType") # Automatic cell type annotation based on cell type signatures.
}
\keyword{Filtering}
\keyword{QC}
\keyword{aware}
\keyword{cell}
\keyword{celltype}
\keyword{filter}
\keyword{type}
