% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSignature.R
\name{ScoreSignatures}
\alias{ScoreSignatures}
\title{Multiple signature scoring on Seurat objects}
\usage{
ScoreSignatures(object, signature.list, assay.use = NA, slot.use = "data")
}
\arguments{
\item{object}{Seurat object.}

\item{signature.list}{The named list of signatures. Each element of the list is a signature (character vector), the names of the elements in the list will be used as signature names.}

\item{assay.use}{The assay within the Seurat object to use for signature scoring. If NA (default), it will use the "imputed" assay if it is present, or the "RNA" assay otherwise.}

\item{slot.use}{The slot to use within the assay to get data for signature scoring. By default, using the "data" assay.}
}
\value{
Seurat object with additional metadata slots containing signature scores.
}
\description{
This function scores multiple signatures at the single cell level on a Seurat object (sum of the expression for each cell and signature).
By default, the score is computed as the average expression of the features as normalized in the "data" slot (i.e. log-normalized typically). Can be overrun by setting the slot.use parameter.
Any feature can be set as a negative marker in the scoring by prefixing it with a minus sign, e.g. "-EPCAM".
If imputed data is available (assay "imputed"), the imputed data will be used. Otherwise, data from the RNA assay will be used. This can be overrun by setting the assay paramter.
}
\examples{
signature.list <- list(EpithelialSignature=c("EPCAM","CDH1","ITGA6"), FibroSignature=c("VIM","COL1A1","ITGA1","-EPCAM"))
MySeuratObject <- Impute(MySeuratObject,signature.list)
MySeuratObject <- ScoreSignatures(MySeuratObject,signature.list)
head(MySeuratObject$EpithelialSignature)
head(MySeuratObject$FibroSignature)
FeaturePlot(MySeuratObject,c("EpithelialSignature","FibroSignature"))
}
\keyword{Multiple}
\keyword{Scoring}
\keyword{Signature}
