% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturePlotQC.R
\name{SignaturePlotQC}
\alias{SignaturePlotQC}
\title{QC plot with signatures highlight}
\usage{
SignaturePlotQC(
  object,
  sign.names,
  x = "nFeature_RNA",
  y = "percent.mito",
  log.scale = TRUE,
  ncol = NA
)
}
\arguments{
\item{object}{Seurat object. Must contain nFeature_RNA and percent.mito metadata columns, or the x and y plot parameters must be changed accordingly to plot other QC info otherwise.}

\item{sign.names}{The names of the signatures to highlight in color. Must correspond to the names of metadata columns.}

\item{x}{Which parameter to set to the x axis in the QC plot (Default: nFeature_RNA).}

\item{y}{Which parameter to set to the y axis in the QC plot (Default: percent.mito).}

\item{log.scale}{Whether to plot with log.scale or not (Default: TRUE)}

\item{ncol}{Number of columns (Default = NA, determined from the data)}
}
\value{
A ggplot grid object.
}
\description{
This function makes quality control (QC) plots from a Seurat object, by default percent.mito vs nFeature_RNA, while highlighting signature scores in color.
}
\details{
This is useful when making scRNA-seq from heterogeneous samples: for example immune cells are small and might be discarded as debris from much larger epithelial cancer cells if one is not aware of the differences in QC parameters occuring between various cell types.
}
\examples{
# After MySeuratObject has been log-normalized in order to contain an RNA>data assay>slot and been augmented with a percent.mito metadata column:
# SignatureList should be a list of character vectors, each containing a series of feature/gene names.
MySeuratObject <- Impute(MySeuratObject)
MySeuratObject <- ScoreSignatures(MySeuratObject,SignatureList)
SignaturePlotQC(MySeuratObject,names(SignatureList))
}
\keyword{QC}
\keyword{highlight}
\keyword{plot}
\keyword{signatures}
