% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSimilarityToGenotypes.R
\name{CellSimilarityToGenotypes}
\alias{CellSimilarityToGenotypes}
\title{Compute the similarity of single cells in a Seurat object to the reference genotypes in a genotype object}
\usage{
CellSimilarityToGenotypes(
  seurat,
  genotype,
  assay.name = "similarity",
  prefix = "Similarity-",
  features.use = "all",
  assays = c("VAR", "REF", "ALT"),
  slots = c("data", "data", "data")
)
}
\arguments{
\item{seurat}{A seurat object}

\item{genotype}{A genotype object}

\item{assay.name}{character(1). The name under which the similarities are stored in the seurat object (Default: "similarity").}

\item{prefix}{character(1). A prefix appended to the genotype names to generate the feature names in the similarity assay (Default: "Similarity-").}

\item{features.use}{character(n). The features (i.e. variants) to be used for the similarity calculations. Can be a vector of variants by name, or NA to use informative variants if available, or "all" (Default: "all").}

\item{assays}{character(3). The name of the assays within the Seurat object that contain the variant calls (sparse matrix with values 1,2,3 in vartrix conventions), and counts of reference and alt alleles (Default: c("VAR","REF","ALT")).}

\item{slots.}{The slots to use within the assays above (Default: c("data","data","data")).}
}
\value{
Returns the Seurat object
}
\description{
Computes the fraction of matching variants between single cells in a Seurat object and the reference genotypes in a genotype object. Store this similarity measure in a new assay within the Seurat object.
}
\examples{
MySeuratObject <- CellSimilarityToGenotypes(MySeuratObject,MyGenotypes)
DefaultAssay(MySeuratObject) <- "similarity"
FeaturePlot(MySeuratObject,"Similarity-MyGenotypeName1")
}
\keyword{cell}
\keyword{compute}
\keyword{fraction}
\keyword{genotype}
\keyword{matching}
\keyword{similarity}
\keyword{variants}
