% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{Classify}
\alias{Classify}
\title{Cell classification based on signatures}
\usage{
Classify(object, signatures, expected.values = NA, metadata.name = "celltype")
}
\arguments{
\item{object}{Seurat object. Must contain nFeature_RNA and percent.mito metadata columns, or the x and y plot parameters must be changed accordingly to plot other QC info otherwise.}

\item{signatures}{A vector containing the names of the signatures to highlight in color. Must correspond to names of metadata columns. If a named list of signatures is passed, the names of the list will be used.}

\item{expected.values}{Vector/List of expected values for the signatures, used for normalization of the signatures before classifying cells. If NA (Default), normalized to max. If 1, no normalization. If The vector/list is named and the names match the signature names, the names will be used to match each scale factor to the right signature. If no matching names are given, the expected values are assumed to be given in the same order as the signatures.}

\item{metadata.name}{The name of the new metadata column where cell type annotations are stored (Default: celltype)}
}
\value{
A Seurat object with an additional metadata column containing the cell type annotations and Idents() set to these annotations.
}
\description{
This function performs a simple cell type classification based on signatures, useful for automating the simple first broad classification of cell types before a cell-type aware QC (for example, taking into account that immune cells have less detected genes/transcripts, and fibroblast less mitochondrial content, than epithelial cancer cells).
}
\details{
Each single cell will be associated to the cell type whose signature is maximal on the cell. A list of expected values for the signatures can be given (for example, checking the typical signature score for cells of known identity beforehand) and are used to normalize signature scores before determining the classification. It is strongly recommended to use imputed data to score the cell type signatures (markers), to avoid mis-classification due on dropouts.
}
\examples{
# After MySeuratObject has been log-normalized in order to contain an RNA>data assay>slot and been augmented with a percent.mito metadata column:
# SignatureList should be a list of character vectors, each containing a series of feature/gene names. In this example, the signatures correspond to cell types, and the name of each signature is the name of the corresponding cell type (e.g. Tcell, Bcell, EpithelialCell, Fibroblasts, etc).
MySeuratObject <- Impute(MySeuratObject)
MySeuratObject <- ScoreSignatures(MySeuratObject,SignatureList)
MySeuratObject <- Classify(MySeuratObject,names(SignatureList),"CellType") # Automatic cell type annotation based on cell type signatures.
}
\keyword{Cell}
\keyword{Classifier}
\keyword{celltype}
\keyword{classification}
\keyword{type}
