% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyManual.R
\name{ClassifyManual}
\alias{ClassifyManual}
\title{Manual cell type classification based on signature gating}
\usage{
ClassifyManual(object, gates, metadata.name = "celltype")
}
\arguments{
\item{object}{Seurat object. Must have metadata columns containing signature scores, such as those generated by ScoreSignatures, with names matching any gating filter supplied.}

\item{gates}{A string (chr) of cell type definitions, with cell types delimited by commas (,). The name of the cell type is given first, followed by equal (=) and a series of gates separated by AND symbol (&). Example: "celltype_a = sign_a > value_aa & sign_b < value_ab, celltype_b = sign_b > value_ba & sign_a < value_ab" where sign_a and sign_b are the names of metadata columns and value_xy are numeric values.}

\item{metadata.name}{The name of the new metadata column where cell type annotations are stored (Default: celltype)}
}
\value{
A Seurat object with an additional metadata column containing the cell type annotations and Idents() set to these annotations. Cells which do not validate any gate defined are attributed the celltype "None" and cells which pass several gates are labelled "Multiplet".
}
\description{
This function performs a cell type classification based on signature gating.
}
\details{
If a list of signatures (sign_a, sign_b...) has been scored with ScoreSignature, cell types (celltype_a, celltype_b...) are defined based on gates (value_xy cutoffs) of the form:
"celltype_a = sign_a > value_aa & sign_b < value_ab, celltype_b = sign_b > value_ba & sign_a < value_ab"
}
\examples{
# MySeuratObject should contain an RNA slot with log-normalized data assay.
# In this example cells are classified as T cells (TC) or Epithelial cells (EP) based on manual signatures.
# Note that cells that validate no gate will have cell type "None", and cells which validate several will be classified as "Multiplets".
# It's recommended to first do a signature scatter plot in order to choose threshold values for the gates.
MySeuratObject <- Impute(MySeuratObject)
signature.list <- list(TCsign=c("CD3D","CD3E","CD3G","CD4","CD8A","CD8B","GZMA","IFNG"),
                       EPsign=c("EPCAM","TFF3","CDX2","CLDN3","KRT8","KRT19","KRT18","OLFM4"))
MySeuratObject <- ScoreSignatures(MySeuratObject,signature.list)
gates <- "TC = TCsign > 1.2 & EPsign < 0.2,
          EP = EPsign > 1.1 & TCsign < 0.3"
MySeuratObject <- ClassifyManual(MySeuratObject,gates)
DimPlot(MySeuratObject) # Plot the current Idents, set to "celltype".
}
\keyword{Cell}
\keyword{Classifier}
\keyword{celltype}
\keyword{classification}
\keyword{gates}
\keyword{gating}
\keyword{manual}
\keyword{type}
