% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyManual.R
\name{ClassifyManual}
\alias{ClassifyManual}
\title{Manual cell type classification by applying gates}
\usage{
ClassifyManual(
  object,
  gates,
  metadata.name = "celltype",
  assay = DefaultAssay(object),
  slot = "data",
  restrict.ident = "Default",
  restrict.to = as.character(unique(Idents(object)))
)
}
\arguments{
\item{object}{Seurat object. Must have metadata columns containing signature scores, such as those generated by ScoreSignatures, with names matching any gating filter supplied.}

\item{gates}{charater(1). Gate definitions, with gates for various cell types delimited by commas (,). For each gate, the name of the cell type is given first, followed by equal (=) and a series of gates separated by AND symbol (&). Example: "celltype_a = sign_a > value_aa & sign_b < value_ab, celltype_b = sign_b > value_ba & sign_a < value_ab" where sign_a and sign_b are the names of metadata columns and value_xy are numeric values.}

\item{metadata.name}{The name of the new metadata column where cell type annotations are stored (Default: celltype)}

\item{assay}{character(1). If some signatures used for classification are stored as assay features, which assay should be used in priority (Default: DefaultAssay(object)).}

\item{slot}{character(1). If some signatures used for classification are stored as assay features, which slot should be used (Default: "data").}

\item{restrict.ident}{character(1). The name of a metadata column containing celltype annotations, which will be used to define the subset of cells that should be classified.}

\item{restrict.to}{character(n). Which celltypes (as defined in the restrict.ident metadata column) should be classified.}
}
\value{
A Seurat object with an additional metadata column containing the cell type annotations and Idents() set to these annotations. Cells which do not validate any gate defined are attributed the celltype "None" and cells which pass several gates are labelled "Multiplet".
}
\description{
This function performs a cell type classification by gating.
}
\details{
Based on cell scores (sign_x, which can be signatures from ScoreSignatures, genetic similarities to reference genotypes, normalized hashtag oligo (HTO) values, gene expression, QC parameters etc), cell types (celltype_a, celltype_b...) are defined based on gates (value_xy cutoffs) of the form:

"celltype_a = sign_a > value_aa & sign_b < value_ab,
celltype_b = sign_b > value_ba & sign_a < value_ab \link{...etc...}"

Scores are taken from the metadata in priority, otherwise from assays. If found in several assays, the assay to use should be defined by prefixing the feature name with "assay_".

Since single cell data tends to be noisy and have a significant amount of dropouts, imputation is usually a great combination with gating strategies, in a similar fashion to gaussian filtering needed before thresholding in most image analysis workflows. For this, see Impute() and CrossImpute().
}
\examples{
# MySeuratObject should contain an RNA slot with log-normalized data assay.
# In this example cells are classified as T cells (TC) or Epithelial cells (EP) based on manual signatures.
# Note that cells that validate no gate will have cell type "None", and cells which validate several will be classified as "Multiplets".
# It's recommended to first do a signature scatter plot in order to choose threshold values for the gates.
MySeuratObject <- Impute(MySeuratObject)
signature.list <- list(TCsign=c("CD3D","CD3E","CD3G","CD4","CD8A","CD8B","GZMA","IFNG"),
                       EPsign=c("EPCAM","TFF3","CDX2","CLDN3","KRT8","KRT19","KRT18","OLFM4"))
MySeuratObject <- ScoreSignatures(MySeuratObject,signature.list)
gates <- "TC = TCsign > 1.2 & EPsign < 0.2,
          EP = EPsign > 1.1 & TCsign < 0.3"
MySeuratObject <- ClassifyManual(MySeuratObject,gates)
DimPlot(MySeuratObject) # Plot the current Idents, set to "celltype".
# Subclassify T cells, and store the results in the metadata column "TC_subcelltype" (Could also be used to overwrite the idents of TC in an existing metadata column without altering the identities of other cells).
gates <- "CD8 = CD8A > 0.5 & CD4 < 0.5, CD4 = CD4 > 0.5 & CD8A < 0.5, DoublePositive = CD8A > 0.5 & CD4 > 0.5"
MySeuratObject <- ClassifyManual(MySeuratObject, gates, "TC_subcelltype", restrict.ident="celltype", restrict.to="TC")
}
\keyword{Cell}
\keyword{Classifier}
\keyword{celltype}
\keyword{classification}
\keyword{gates}
\keyword{gating}
\keyword{manual}
\keyword{type}
