% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export10X.R
\name{Export10X}
\alias{Export10X}
\title{Reexport the data from a Seurat object in 10X format}
\usage{
Export10X(
  object,
  dir,
  meta_columns = NA,
  append_reductions = c(),
  gzip = T,
  rows = "features.tsv",
  cols = "barcodes.tsv",
  counts = "matrix.mtx",
  meta = "metadata.tsv"
)
}
\arguments{
\item{object}{Seurat object.}

\item{meta_columns}{character(n). Metadata columns that should be exported from the Seurat Object. NA for all, empty vector c() for none. Default: NA.}

\item{append_reductions}{character(n). Name of the dimensionality reductions that should be included with the metadata. Default: none.}

\item{gzip}{logical(1). Should the exported files be compressed. If compressed, the filenames are appended with the extension ".gz". Default: TRUE}

\item{rows}{character(1). Name of the exported file that lists row names. Default: features.tsv}

\item{cols}{character(1). Name of the exported file that lists column names. Default: barcodes.tsv}

\item{counts}{character(1). Name of the exported file that contains the sparse raw count matrix. Default: matrix.mtx}

\item{meta}{character(1). Name of the exported file that contains the metadata. Default: metadata.tsv}
}
\description{
Exports the counts matrix, features and barcodes from a Seurat object in a 10X-like format, with an additional metadata matrix in tsv format, that can be augmented dimensionality reduction coordinates. The exported data is in standard mtx and tsv formats, which facilitates sharing and reuse, and can be re-imported in a Seurat object with Import10X.
}
\examples{
Export10X(SeuratObject) # By default keeping all the metadata columns, and no dimensionality reductions.
Export10X(SeuratObject,c("nFeature_RNA","mito.content"),c("pca","umap"),gzip=FALSE) # Include only two custom metadata columns, two dimensionality reductions, and do not compress the matrices.
}
\keyword{10X}
\keyword{Reexport}
\keyword{export}
