% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindInformativeVariants.R
\name{FindInformativeVariants}
\alias{FindInformativeVariants}
\title{Find informative variants}
\usage{
FindInformativeVariants(genotype, seurat, n.variants = 10000, assay = "VAR")
}
\arguments{
\item{genotype}{A genotype object.}

\item{seurat}{A seurat object.}

\item{n.variants}{numeric(1). The number of most informative variants stored in the @informative_variants slot (Default: 10000).}

\item{assay}{character(1). The name of the assay which stores the single cell variant calls data within the Seurat object, in vartrix formatting (sparse matrix with values 1-2-3 for ref-alt-het).}
}
\value{
Returns the updated genotype object
}
\description{
Computes the fraction of detection of variants at the single cell level, and entropy associated with the variants, based on a genotype and a Seurat objects. Stores this information in the genotype object metadata, and compute the most informative variants. Returns the updated genotype object.
}
\details{
FindInformativeVariants populates the following metadata columns and slots in the genotype object:

$excess_entropy (see below),

$coverage (number of cells with data for the variant),

$coverage_frac (fraction of the cell with data for the variant),

@variants_by_coverage (variants sorted from top to least coverage),

@variants_by_information (variants sorted from top to least excess entropy),

@informative_variants (most informative variants, to be used similarly to VariableFeatures for continuous data within Seurat Objects for further analysis).

The amount of information provided by a variant at the single cell level is computed as the Shannon entropy, sum(-p*log2(p)), minus the minimal entropy contributed by the mere coverage of the variant, i.e. the entropy value if the data had just the two levels nodata and data. This is analogous to the excess variance for continuous data, translated to discrete variant data. This quantity is referred to as excess_entropy in the genotype metadata.
}
\examples{
MyGenotype <- FindInformativeVariants(MyGenotype, MySeuratObject)
}
\keyword{entropy}
\keyword{excess}
\keyword{features}
\keyword{genotype}
\keyword{informative}
\keyword{most}
\keyword{variable}
\keyword{variants}
