% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{Impute}
\alias{Impute}
\title{MAGIC imputation}
\usage{
Impute(object, features = NA, npca = 100, knn = 3, t = 2, n.jobs = 6)
}
\arguments{
\item{object}{Seurat object. Must contain an RNA assay with a data slot that will be used for imputation.}

\item{features}{A subset of features on which to run the imputation. NA (default) uses VariableFeatures(object).}

\item{npca}{Number of principal components to use for imputation.}

\item{knn}{Number of nearest neighbors on which to compute bandiwth imputation.}

\item{t}{Diffusion parameter passed to MAGIC.}

\item{n.jobs}{Number of threads on which to run the imputation. Defaults to 6.}
}
\value{
Seurat object with an additional slot (accessed as object[\link{"imputed"}]) containing MAGIC imputed data.
}
\description{
This function is a wrapper of the MAGIC imputation function to easily impute Seurat objects containing scRNA-seq data, with more reasonable default parameters.
}
\examples{
MySeuratObject <- Impute(MySeuratObject)
MySeuratObject[["imputed"]]
}
\keyword{MAGIC}
\keyword{Rmagic}
\keyword{imputation}
