% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeHTO.R
\name{NormalizeHTO}
\alias{NormalizeHTO}
\title{Normalize hashtag oligo data}
\usage{
NormalizeHTO(
  object,
  scale.HTO = "stdev",
  normalize.cells = 100,
  assay = "HTO",
  q = 0.9,
  imputed.assay = "imputed.HTO"
)
}
\arguments{
\item{object}{Seurat object.}

\item{scale.HTO}{character(1) or logical(1). "stdev"/TRUE/T to scale each HTO to standard deviation across cells, No/no/FALSE/F for no scaling, "quantile" to scale to a quantile of values, "background" to scale to the average background value, and "imputed_max" to scale to the max after smoothing by imputation to avoid outliers. Default: "stdev".}

\item{normalize.cells}{numeric(1). The value to which each cell should be normalized. 0 for no cell-normalization. Default: 100.}

\item{assay}{character(1). The name of the assay which should be normalized. Default: "HTO".}

\item{q}{numeric(1). Only used if scale.HTO="quantile". The quantile used for normalization.}

\item{imputed.assay}{character(1). Only used if scale.HTO="imputed_max". The name of the assay that contains imputed HTO data.}
}
\value{
Returns a Seurat object with normalized HTO data in the data slot.
}
\description{
Normalizes hashtag oligo (HTO) data. By default, first scales each HTO to its standard deviation to account for difference of intensity for the various labels, then scales each cell so that HTO are reported as a percentage. Alternatively, can scale the HTOs to a quantile of their values, to the background value, or to their max after imputation.
}
\details{
When normalizing to the background value, a k-means clustering with k=2 is used to define a positive and a negative population, and the average of the negative population is used for scaling (i.e. set to 1).
}
\examples{
MySeuratObject <- NormalizeHTO(MySeuratObject)
}
\keyword{HTO}
\keyword{normalization}
