% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVartrix.R
\name{ReadVartrix}
\alias{ReadVartrix}
\title{Load vartrix data}
\usage{
ReadVartrix(ref_matrix, alt_matrix, cell_barcodes, variants, tolerance_pct = 5)
}
\arguments{
\item{ref_matrix}{character(1), the file name for the reference allele count matrix.}

\item{alt_matrix}{character(1), the file name for the alternate allele count matrix.}

\item{cell_barcodes}{character vector, the list of cell barcodes for which vartrix was run, used as column names on the matrices.}

\item{variants}{character vector, the list of variants for which vartrix was run, used as row names on the matrices.}

\item{tolerance_pct}{numeric(1), the percentage of alt or ref counts not in agreement with the rest on a unique (cell,variant) tolerated without affecting the call. Default: 5\%.}
}
\value{
A list of sparse matrices (dgCmatrix), with names ref, alt, freq, and consensus, variants as rows and cells/barcodes as columns.
}
\description{
Read the vartrix ref and alt matrices, and computes a frequency and a consensus matrix. Returns all of them as a list of sparse matrices.
}
\details{
The frequency sparse matrix is offset by 1 in order to distinguish zero-frequency (value 1) from missing data (value 0, not stored in the sparse matrix).

The consensus matrix makes a call for no data (value 0, not stored in the sparse matrix), ref/ref (value 1), alt/alt (value 2) or alt/ref (value 3) genotypes. Since in scRNA-seq data, there is likely some extracellular RNA from dead cells or debris slightly contaminating the reads of other cells, a tolerance can be set above 0. For example, at a tolerance of 5\%, up to 1 read in 20 can differ from the rest without making the call switch to heterozygous. The behavior of the vartrix consensus matrix calculation corresponds to a tolerance of 0.

Equivalence between vartrix, human, and vcf genotype naming conventions:

"0" for "no call", "./."

"1" for "ref/ref", "0/0"

"2" for "alt/alt", "1/1"

"3" for "ref/alt", "0/1"
}
\examples{
vartrix_matrices <- ReadVartrix("MyVartrixRef.mtx.gz","MyVartrixAlt.mtx.gz",MyCellBarcodes,MyVariants)
}
\keyword{alt}
\keyword{consensus}
\keyword{freq}
\keyword{frequency}
\keyword{matrix}
\keyword{ref}
\keyword{variants}
\keyword{vartrix}
